+++ webviz_config/_deployment/radix.py
@@ -1,3 +1,4 @@
+import sys
 import json
 import time
 import secrets
@@ -99,10 +100,11 @@ def radix_initial_deployment(github_slug: str, build_directory: pathlib.Path) ->
 
     progress_bar.write("✓ Uploading Webviz file resources to Azure storage container.")
     azure_cli.storage_container_upload_folder(
-        azure_configuration_values["storage_account_name"],
-        azure_configuration_values["storage_container_name"],
-        ".",
-        build_directory / "resources",
+        subscription=azure_configuration_values["subscription"],
+        resource_group=azure_configuration_values["resource_group"],
+        storage_name=azure_configuration_values["storage_account_name"],
+        container_name=azure_configuration_values["storage_container_name"],
+        source_folder=build_directory / "resources",
     )
     progress_bar.update()
 
@@ -258,21 +260,21 @@ def radix_redeploy(github_slug: str, build_directory: pathlib.Path) -> None:
     github_cli.upload_directory(github_slug, build_directory, commit_message="Update")
 
     azure_cli.storage_container_upload_folder(
-        deploy_settings["azure"]["storage_account_name"],
-        deploy_settings["azure"]["storage_container_name"],
-        ".",
-        build_directory / "resources",
+        subscription=deploy_settings["azure"]["subscription"],
+        resource_group=deploy_settings["azure"]["resource_group"],
+        storage_name=deploy_settings["azure"]["storage_account_name"],
+        container_name=deploy_settings["azure"]["storage_container_name"],
+        source_folder=build_directory / "resources",
     )
 
 
 def main_radix_deployment(args: argparse.Namespace) -> None:
 
+    if sys.version_info < (3, 8):
+        raise RuntimeError("Radix deployment workflow requires at least Python 3.8")
+
     if not args.portable_app.is_dir():
         raise ValueError(f"{args.portable_app} is not a directory.")
-
-    if not azure_cli.logged_in():
-        print("You are not logged in with Azure CLI. Follow instructions below.")
-        azure_cli.log_in()
 
     if args.initial_deploy:
         radix_initial_deployment(args.github_slug, args.portable_app)